refactor: Plan 엔티티 FSD 구조 변경
