+++ src/entities/plan/api/endpoint.ts
@@ -1,0 +1,5 @@
+export const PLAN_URL = {
+  base: '/plans',
+  plans: () => `${PLAN_URL.base}`,
+  detail: (id: string) => `${PLAN_URL.base}/${id}`,
+}
