+++ src/entities/plan/api/queryKey.ts
@@ -1,0 +1,4 @@
+export const PLAN_QUERY_KEY = {
+  all: ['plans'] as const,
+  detail: (id: string) => ['plan', id] as const,
+}
