+++ src/views/list-plan/index.tsx
@@ -4,9 +4,8 @@ import FloatingWriteButton from '@/src/widgets/floating-write-btn'
 import ProfileImage from '@/src/shared/ui/ProfileImage'
 import Spacer from '@/src/shared/ui/Spacer'
 import BlankTooltip from '@/src/features/plan/blank-tooltip'
-import { PlanType } from '@/src/entities/plan/type'
+import { PlanType, useGetPlans } from '@/src/entities/plan'
 import CardPlan from '@/src/features/plan/card-plan'
-import { useGetPlans } from '@/src/entities/plan/query'
 import { getLoginUrl } from '@/src/entities/login/api'
 import { useRouter } from 'next/navigation'
 import { useState } from 'react'
