+++ src/widgets/course-plan-detail-layout/index.tsx
@@ -7,7 +7,7 @@ import type { CourseType } from '@/src/entities/course'
 import { CATEGORY } from '@/src/entities/course'
 import { OptionHeader } from '@/src/widgets/header'
 import useUserStore from '@/src/shared/store/userStore'
-import { PlanType } from '@/src/entities/plan/type'
+import { PlanType } from '@/src/entities/plan'
 import { Share2, X } from 'lucide-react'
 import { useRouter } from 'next/navigation'
 import { useMemo, useState } from 'react'
