+++ src/widgets/header/index.tsx
@@ -9,7 +9,7 @@ import React from 'react'
 import setting from '@/src/assets/images/setting.png'
 import useUserStore from '@/src/shared/store/userStore'
 import { useQueryClient } from '@tanstack/react-query'
-import { useDeletePlan } from '@/src/entities/plan/query'
+import { useDeletePlan } from '@/src/entities/plan'
 import OptionDropbox from '@/src/shared/ui/OptionDropbox'
 import {
   useDeleteCourse,
